ae: move agent/ae to agent/local/ae
This package is used exclusively as a way of syncing local state. By moving it under local/ae
that relationship becomes more explicit.
